chore: fix circular dependency issues
