+++ src/entityOperations/apiRequests.ts
@@ -4,11 +4,9 @@ import type { SagaIterator } from "redux-saga";
 import { call, delay } from "redux-saga/effects";
 
 import { credentialsByToolNameSelector } from "~/modules/credentials/credentialsSelectors";
-import { getStore } from "~/redux/configureStore";
 import { ToolName } from "~/typeDefs";
 import {
   getApiUrl,
-  isTestingMode,
   isUseLocalApi,
   TogglApiContext,
 } from "~/utilities/environment";
@@ -135,14 +133,6 @@ async function fetchWithRetries<TResponse>(
  * and stable API.
  */
 export function getApiDelayForTool(toolName: ToolName): number {
-  try {
-    if (isTestingMode()) {
-      return 0;
-    }
-  } catch {
-    return 0;
-  }
-
   switch (toolName) {
     case ToolName.Clockify:
       return isUseLocalApi() ? 0 : 1000 / 8;
@@ -158,7 +148,7 @@ export function getApiDelayForTool(toolName: ToolName): number {
 async function fetchFromApi<T>(url: string, config: RequestInit): Promise<T> {
   const { toolName, context, endpoint } = extrapolateFromUrl(url);
 
-  const state = getStore().getState();
+  const state = window.store.getState();
 
   const credentialsByToolName = credentialsByToolNameSelector(state);
