+++ src/modules/allEntities/__tests__/allEntitiesSaga.test.ts
@@ -13,6 +13,7 @@ import { EntityGroup } from "~/typeDefs";
 import { allEntitiesSaga } from "../allEntitiesSaga";
 
 jest.mock("~/modules/app/workflowStep");
+jest.mock("~/utilities/environment");
 
 describe("within allEntitiesSaga", () => {
   test("the createAllEntitiesSaga dispatches the correct actions", () => {
